db fixes
